Updated the naming of certain components and added new ones.
I updated the names of the laser components since David needed me to update them for the battery laser pointers he ordered for the Germany Trip. So I made some new holders and also made a cutting assistance tool to make cutting the rods and motors safer and easier on a Bandsaw.